Introduce necessary deprecation notices
