+++ projects/plugins/jetpack/class.jetpack-network.php
@@ -5,10 +5,8 @@
  * @package automattic/jetpack
  */
 
-use Automattic\Jetpack\Connection\Client;
 use Automattic\Jetpack\Connection\Manager;
-use Automattic\Jetpack\Connection\Utils as Connection_Utils;
-use Automattic\Jetpack\Constants;
+use Automattic\Jetpack\Connection\Tokens;
 use Automattic\Jetpack\Status;
 
 /**
@@ -494,7 +492,7 @@ class Jetpack_Network {
 	 */
 	public function filter_register_user_token( $token ) {
 		$is_master_user = ! Jetpack::is_active();
-		Connection_Utils::update_user_token(
+		Tokens::update_user_token(
 			get_current_user_id(),
 			sprintf( '%s.%d', $token->secret, get_current_user_id() ),
 			$is_master_user
